+++ packages/mars/src/shims-dom.d.ts
@@ -1,0 +1,3 @@
+interface CSSStyleDeclaration {
+  [prop: string]: any;
+}
